接口调通

+++ src/main/java/com/dx/facade/game/resp/GameOrderTableDetailSumResp.java
@@ -45,7 +45,7 @@ public class GameOrderTableDetailSumResp implements Serializable {
     /**
      * 牌桌手数
      */
-    private int handCount;
+    private Long handCount;
     /**
      * 牌桌流水
      */
